spider_system/requirements.txt
fastapi
uvicorn
scrapy
elasticsearch
elasticsearch-dsl
transformers
sentence-transformers
torch
scrapy-playwright
playwright
fake_useragent
textblob
python-dotenv
psutil
colorama
python-dotenv